changed GroupHome to SingleGroup
