+++ client/components/GroupHome.js
@@ -1,55 +1,55 @@
-import React, { Component } from 'react';
-import TextField from 'material-ui/TextField';
-import AutoComplete from 'material-ui/AutoComplete';
-import Chip from 'material-ui/Chip';
-import Avatar from 'material-ui/Avatar';
-import RaisedButton from 'material-ui/RaisedButton';
-import { getGroupsByUserId, getGroupById, addGroup, updateGroup, deleteGroup } from '../crud/group';
-import Paper from 'material-ui/Paper';
-import UsersList from './UsersList';
-import DocsList from './DocsList';
+// import React, { Component } from 'react';
+// import TextField from 'material-ui/TextField';
+// import AutoComplete from 'material-ui/AutoComplete';
+// import Chip from 'material-ui/Chip';
+// import Avatar from 'material-ui/Avatar';
+// import RaisedButton from 'material-ui/RaisedButton';
+// import { getGroupsByUserId, getGroupById, addGroup, updateGroup, deleteGroup } from '../crud/group';
+// import Paper from 'material-ui/Paper';
+// import UsersList from './UsersList';
+// import DocsList from './DocsList';
 
 
-const styles = theme => ({
-  container: {
-    display: 'grid',
-    gridTemplateColumns: 'repeat(12, 1fr)',
-    gridGap: `${theme.spacing.unit * 3}px`,
-  },
-  paper: {
-    padding: theme.spacing.unit,
-    textAlign: 'center',
-    color: theme.palette.text.secondary,
-    whiteSpace: 'nowrap',
-    marginBottom: theme.spacing.unit,
-  },
-  divider: {
-    margin: `${theme.spacing.unit * 2}px 0`,
-  },
-});
-
-
-
-class GroupHome extends Component {
-  // constructor(){
-
-  // }
+// const styles = theme => ({
+//   container: {
+//     display: 'grid',
+//     gridTemplateColumns: 'repeat(12, 1fr)',
+//     gridGap: `${theme.spacing.unit * 3}px`,
+//   },
+//   paper: {
+//     padding: theme.spacing.unit,
+//     textAlign: 'center',
+//     color: theme.palette.text.secondary,
+//     whiteSpace: 'nowrap',
+//     marginBottom: theme.spacing.unit,
+//   },
+//   divider: {
+//     margin: `${theme.spacing.unit * 2}px 0`,
+//   },
+// });
 
 
 
-  render() {
+// class GroupHome extends Component {
+//   // constructor(){
 
-    return (
-      <div>
-        <Paper className={styles.paper}>
-          <UsersList groupId={'Group 2'}/>
-        </Paper>
-        <Paper className={styles.paper}>
-          <DocsList groupId={'Group 2'}/>
-        </Paper>
-      </div>
-    )
-  }
-}
+//   // }
 
-export default GroupHome;
+
+
+//   render() {
+
+//     return (
+//       <div>
+//         <Paper className={styles.paper}>
+//           <UsersList groupId={'Group 2'}/>
+//         </Paper>
+//         <Paper className={styles.paper}>
+//           <DocsList groupId={'Group 2'}/>
+//         </Paper>
+//       </div>
+//     )
+//   }
+// }
+
+// export default GroupHome;
