+++ client/components/navbar.js
@@ -23,6 +23,7 @@ const Navbar = ({ handleClick, isLoggedIn }) => (
           <Link to="/signup">Sign Up</Link>
           <Link to="/testSnippet">Example Snippet</Link>
           <Link to="/CreateGroup">CreateGroup</Link>
+          <Link to="/groups/Group 2">Show SingleGroup (Just for demonstrational purpose!!)</Link>
         </div>
       )}
     </nav>
