+++ client/routes.js
@@ -23,24 +23,19 @@ const Routes = () => (
     <Switch>
       <Route path="/login" component={Login} />
       <Route path="/signup" component={Signup} />
-    </Switch>
 
-    <Switch>
+
       <Route exact path="/groups" component={AllGroups} /> {/* all groups ever */}
       <Route exact path="/groups/:groupId" component={SingleGroup} /> {/* this group */}
       <Route path="/groups/:groupId/users" component={AllUsers} /> {/* filtered by who's in this group */}
       <Route path="/groups/:groupId/notebooks" component={AllNotebooks} /> {/* all notebooks for this group */}
       <Route path="/groups/new" component={CreateGroup} />
-    </Switch>
 
-    <Switch>
       <Route exact path="/users" component={AllUsers} /> {/* all users ever */}
       <Route exact path="/users/:userId" component={SingleUser} /> {/* this user */}
       <Route path="/users/:userId/groups" component={AllGroups} /> {/* filtered by who's in this user */}
       <Route path="/users/:userId/notebooks" component={AllNotebooks} /> {/* all notebooks for this user */}
-    </Switch>
 
-    <Switch>
       <Route exact path="/notebooks" component={AllNotebooks} /> {/* created by anyone, ever */}
       <Route path="/notebooks/:notebookId" component={SingleNotebook} /> {/* this notebook */}
     </Switch>
